Move all the server model code out of server.ts, letting server.ts focus on, well, serving.
